small changes to publ

+++ publisher.c
@@ -17,9 +17,7 @@
 #define TEMPFILENAME "publ_db2.txt"
 
 typedef struct Publisher{
-    char server_message[MSGSIZ];
-    char client_message[MSGSIZ];
-    int socket_desc;
+    int socket_fd;
     struct sockaddr_in server_addr;
     MessageHeader msg_header;
     int db_fd;
@@ -31,13 +29,8 @@ Publisher* initServer(char* _addr, int _port){
         perror("Error allocating memory");
         exit(EXIT_FAILURE);
     }
-
-    memset(_pub->server_message,'\0',sizeof(_pub->server_message));
-    memset(_pub->client_message,'\0',sizeof(_pub->client_message));
-    
-    _pub->socket_desc = socket(AF_INET, SOCK_STREAM, 0);
- 
-    if(_pub->socket_desc < 0){
+    _pub->socket_fd = socket(AF_INET, SOCK_STREAM, 0);
+    if(_pub->socket_fd < 0){
         perror("Unable to create socket");
         exit(EXIT_FAILURE);
     }
@@ -52,8 +45,7 @@ Publisher* initServer(char* _addr, int _port){
         perror("Error opening messages file");
         exit(EXIT_FAILURE);
     }
-
-    if(connect(_pub->socket_desc, (struct sockaddr*)&(_pub->server_addr), sizeof(_pub->server_addr)) < 0){
+    if(connect(_pub->socket_fd, (struct sockaddr*)&(_pub->server_addr), sizeof(_pub->server_addr)) < 0){
         perror("Unable to connect");
         //exit(EXIT_FAILURE);
     }
@@ -62,23 +54,27 @@ Publisher* initServer(char* _addr, int _port){
     return _pub;
 }
 
-void sendMessage(Publisher* _pub){
+void sendMessage(Publisher* _pub){ 
+    //placeholder function
+    char client_message[MSGSIZ], server_message[MSGSIZ];
+    memset(client_message, '\0', MSGSIZ);
+    memset(server_message, '\0', MSGSIZ);
+
     printf("Enter message: ");
-    fgets(_pub->client_message, MSGSIZ, stdin);
+    fgets(client_message, MSGSIZ, stdin);
  
-    if(send(_pub->socket_desc, _pub->client_message, strlen(_pub->client_message), 0) < 0){
+    if(send(_pub->socket_fd, client_message, strlen(client_message), 0) < 0){
         perror("Unable to send message");
         exit(EXIT_FAILURE);
     }
-     if(recv(_pub->socket_desc, _pub->server_message, sizeof(_pub->server_message), 0) < 0){
+     if(recv(_pub->socket_fd, server_message, sizeof(server_message), 0) < 0){
         perror("Error while receiving server's msg\n");
         exit(EXIT_FAILURE);
     }
-    printf("Server's response: %s\n",_pub->server_message);
+    printf("Server's response: %s\n",server_message);
 }
 
-void fetchMessages(Publisher* _pub){
-    setbuf(stdout, NULL);
+void fetchAndSend(Publisher* _pub){
     if(lseek(_pub->db_fd, 0, SEEK_END) == 0){
         printf("Empty database");
         exit(EXIT_FAILURE);
@@ -92,9 +88,10 @@ void fetchMessages(Publisher* _pub){
     char buf[BUFSIZ];
     char *msg_buf = NULL;
     char header_buf[sizeof(MessageHeader)];
-    printf("hi, %lu\n", sizeof(MessageHeader));
     int ret;
+    bool was_modified = 0;
     while((ret = read(_pub->db_fd, header_buf, sizeof(MessageHeader))) > 0){
+        was_modified = 1;
         printf("hi from while\n");
         memcpy(&(_pub->msg_header), header_buf, sizeof(MessageHeader));
         printf("type:%i\npriority:%i\npubID:%i\nlenght:%i\n",
@@ -110,32 +107,33 @@ void fetchMessages(Publisher* _pub){
         read(_pub->db_fd, msg_buf, _pub->msg_header.len);
         printf("%s\n", msg_buf);
         //send to server
-
+        
 
     }
-    //exited while & messages unread remained
-    while((ret = read(_pub->db_fd, buf, BUFSIZ)) > 0){
-        write(temp_fd, buf, ret);
+    if(was_modified){
+        while((ret = read(_pub->db_fd, buf, BUFSIZ)) > 0){
+            write(temp_fd, buf, ret);
+        }
+        close(_pub->db_fd);
+        _pub->db_fd = temp_fd;
+        unlink(MSGFILENAME);
+        rename(TEMPFILENAME, MSGFILENAME);
     }
-    close(_pub->db_fd);
-    _pub->db_fd = temp_fd;
-    unlink(MSGFILENAME);
-    rename(TEMPFILENAME, MSGFILENAME);
-
 
 }
 void closePubClient(Publisher* _pub){
     close(_pub->db_fd);
-    close(_pub->socket_desc);
+    close(_pub->socket_fd);
     free(_pub);
 }
 
 
 int main(void)
 {
+    setbuf(stdout, NULL);
     Publisher* publisher_client = initServer(IPADDR, PORT);
  
-    fetchMessages(publisher_client);
+    fetchAndSend(publisher_client);
     closePubClient(publisher_client);
     
     return 0;
